requirements.txt
django>=5.0
mysqlclient
Pillow
django-allauth
requests
pyjwt
cryptography
